Split simple transfer transaction tests
